# Docs/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AUSTA Cockpit API
  description: |
    Medical Audit Platform API with AI-powered decision support.
    
    This API provides endpoints for managing medical authorization cases, 
    AI-assisted analysis, audit decisions, and comprehensive analytics.
    
    ## Authentication
    
    The API uses JWT tokens for authentication. Obtain a token through the 
    `/auth/login` endpoint and include it in the Authorization header:
    
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    
    API calls are rate limited to ensure fair usage:
    - Anonymous: 100 requests/hour
    - Authenticated: 1000 requests/hour
    - Premium: 10000 requests/hour
    
    ## Environments
    
    - Development: `http://localhost:3000/api/v1`
    - Staging: `https://staging-api.austa.com.br/v1`
    - Production: `https://api.austa.com.br/v1`
    
  version: 1.0.0
  contact:
    name: AUSTA Support
    email: suporte@austa.com.br
    url: https://docs.austa.com.br
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://staging-api.austa.com.br/v1
    description: Staging server
  - url: https://api.austa.com.br/v1
    description: Production server
...
paths:
...
  /cases:
    get:
      tags:
        - Cases
      summary: List cases
      description: Get paginated list of medical authorization cases
      operationId: listCases
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - name: status
          in: query
          description: Filter by case status
          schema:
            type: string
            enum: [pending, in_review, approved, denied, partial]
        - name: priority
          in: query
          description: Filter by priority
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: assignedTo
          in: query
          description: Filter by assigned auditor ID
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          description: Search in case description and patient info
          schema:
            type: string
      responses:
        '200':
          description: Cases retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Case'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Cases
      summary: Create new case
      description: Submit a new medical authorization case for review
      operationId: createCase
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaseRequest'
      responses:
        '201':
          description: Case created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /cases/{caseId}:
    get:
      tags:
        - Cases
      summary: Get case details
      description: Retrieve detailed information about a specific case
      operationId: getCase
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CaseIdParam'
      responses:
        '200':
          description: Case details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Cases
      summary: Update case
      description: Update case information or status
      operationId: updateCase
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CaseIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCaseRequest'
      responses:
        '200':
          description: Case updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /cases/{caseId}/assign:
    post:
      tags:
        - Cases
      summary: Assign case to auditor
      description: Assign or reassign a case to an auditor
      operationId: assignCase
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CaseIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - auditorId
              properties:
                auditorId:
                  type: string
                  format: uuid
                  description: ID of the auditor to assign
      responses:
        '200':
          description: Case assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  parameters:
    CaseIdParam:
      name: caseId
      in: path
      required: true
      description: Unique case identifier
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      description: Sort field and order (e.g., "createdAt:desc")
      schema:
        type: string
        pattern: '^[a-zA-Z]+:(asc|desc)$'
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [auditor, supervisor, admin, viewer]
        avatar:
          type: string
          format: uri
        mfaEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
...
    Case:
      type: object
      properties:
        id:
          type: string
          format: uuid
        caseNumber:
          type: string
          example: "AUT-2024-7834"
        patientId:
          type: string
          format: uuid
        patientName:
          type: string
        procedureCode:
          type: string
        procedureDescription:
          type: string
        requestedValue:
          type: number
          format: float
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [pending, in_review, approved, denied, partial]
        assignedTo:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
    CaseDetail:
      allOf:
        - $ref: '#/components/schemas/Case'
        - type: object
          properties:
            medicalHistory:
              type: array
              items:
                type: object
                properties:
                  date:
                    type: string
                    format: date
                  diagnosis:
                    type: string
                  treatment:
                    type: string
            documents:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  type:
                    type: string
                    enum: [prescription, exam, report, authorization]
                  url:
                    type: string
                    format: uri
                  uploadedAt:
                    type: string
                    format: date-time
            timeline:
              type: array
              items:
                type: object
                properties:
                  timestamp:
                    type: string
                    format: date-time
                  event:
                    type: string
                  user:
                    type: string
                  details:
                    type: string
    CreateCaseRequest:
      type: object
      required:
        - patientId
        - procedureCode
        - requestedValue
      properties:
        patientId:
          type: string
          format: uuid
        procedureCode:
          type: string
        procedureDescription:
          type: string
        requestedValue:
          type: number
          format: float
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        documents:
          type: array
          items:
            type: string
            format: uuid
            description: Document IDs to attach
        notes:
          type: string
          description: Additional notes or context
    UpdateCaseRequest:
      type: object
      properties:
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [pending, in_review, approved, denied, partial]
        notes:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            timestamp:
              type: string
              format: date-time
            path:
              type: string
            requestId:
              type: string
              format: uuid
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
  responses:
    BadRequest:
      description: Bad request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnprocessableEntity:
      description: Unprocessable entity - Validation errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...